+++ examples/flow.php
@@ -4,28 +4,30 @@ declare(strict_types=1);
 
 require __DIR__.'/../vendor/autoload.php';
 
-use function Amp\delay;
+use function Amp\delay as amp_delay;
+use function React\Async\delay as react_delay;
+
 use Flow\Driver\AmpDriver;
 use Flow\Driver\ReactDriver;
 use Flow\Driver\SwooleDriver;
+use Flow\Flow\Flow;
 use Flow\Ip;
 use Flow\IpStrategy\MaxIpStrategy;
-use Flow\Flow\Flow;
 use Swoole\Coroutine;
 
-$randomDriver = random_int(1, 5);
+$randomDriver = random_int(1, 3);
 
 if (1 === $randomDriver) {
     printf("Use AmpDriver\n");
 
     $driver = new AmpDriver();
 
-    $job1 = static function (object $data): Generator {
+    $job1 = static function (object $data): void {
         printf("*. #%d - Job 1 : Calculating %d + %d\n", $data['id'], $data['number'], $data['number']);
 
-        // simulating calculating some "light" operation from 100 to 900 milliseconds as async generator
+        // simulating calculating some "light" operation from 100 to 900 milliseconds
         $delay = random_int(1, 9) * 100;
-        yield delay($delay);
+        amp_delay($delay / 1000);
         $result = $data['number'];
         $result += $result;
 
@@ -45,7 +47,7 @@ if (1 === $randomDriver) {
     $job1 = static function (object $data): void {
         printf("*. #%d - Job 1 : Calculating %d + %d\n", $data['id'], $data['number'], $data['number']);
 
-        // simulating calculating some "light" operation from 100 to 900 milliseconds as async generator
+        // simulating calculating some "light" operation from 100 to 900 milliseconds
         $delay = random_int(1, 9) * 100;
         Coroutine::sleep($delay / 1000);
         $result = $data['number'];
@@ -67,7 +69,9 @@ if (1 === $randomDriver) {
     $job1 = static function (object $data): void {
         printf("*. #%d - Job 1 : Calculating %d + %d\n", $data['id'], $data['number'], $data['number']);
 
-        // simulating calculating some "light"
+        // simulating calculating some "light" operation from 100 to 900 milliseconds
+        $delay = random_int(1, 9) * 100;
+        react_delay($delay / 1000);
         $result = $data['number'];
         $result += $result;
 
@@ -75,7 +79,7 @@ if (1 === $randomDriver) {
             throw new Error('Failure when processing "Job1"');
         }
 
-        printf("*. #%d - Job 1 : Result for %d + %d = %d\n", $data['id'], $data['number'], $data['number'], $result);
+        printf("*. #%d - Job 1 : Result for %d + %d = %d and took %d milliseconds\n", $data['id'], $data['number'], $data['number'], $result, $delay);
 
         $data['number'] = $result;
     };
@@ -103,13 +107,12 @@ $errorJob = static function (object $data, Throwable $exception): void {
     $data['number'] = null;
 };
 
-
 $flow = (new Flow($job1, $errorJob, new MaxIpStrategy(2), $driver))
     ->fn(new Flow($job2, $errorJob, new MaxIpStrategy(2), $driver));
 
 $ipPool = new SplObjectStorage();
 
-for ($i = 1; $i <= 5; ++$i) {
+for ($i = 1; $i <= 5; $i++) {
     $ip = new Ip(new ArrayObject(['id' => $i, 'number' => $i]));
     $ipPool->offsetSet($ip, true);
     $flow($ip, fn ($ip) => $ipPool->offsetUnset($ip));