+++ src/Flow/Flow.php
@@ -7,10 +7,10 @@ namespace Flow\Flow;
 use Closure;
 use Flow\Driver\AmpDriver;
 use Flow\DriverInterface;
+use Flow\FlowInterface;
 use Flow\Ip;
 use Flow\IpStrategy\LinearIpStrategy;
 use Flow\IpStrategyInterface;
-use Flow\FlowInterface;
 use SplObjectStorage;
 use Throwable;
 
@@ -57,7 +57,7 @@ class Flow implements FlowInterface
     {
         $ip = $this->ipStrategy->pop();
         if (!$ip) {
-            return ;
+            return;
         }
 
         $callback = $this->callbacks->offsetGet($ip);
@@ -67,19 +67,19 @@ class Flow implements FlowInterface
         foreach ($this->jobs as $i => $job) {
             $this->driver->coroutine($job, function (Throwable $exception = null) use ($ip, &$count, $i, $callback) {
                 $count--;
-                if($count === 0 || $exception !== null) {
+                if ($count === 0 || $exception !== null) {
                     $count = 0;
                     $this->ipStrategy->done($ip);
                     $this->nextIpJob();
 
-                    if($exception) {
-                        if(isset($this->errorJobs[$i])) {
+                    if ($exception) {
+                        if (isset($this->errorJobs[$i])) {
                             $this->errorJobs[$i]($ip->data, $exception);
                         } else {
                             throw $exception;
                         }
                     }
-                    
+
                     if ($this->fnFlow) {
                         ($this->fnFlow)($ip, $callback);
                     } else {
@@ -99,7 +99,7 @@ class Flow implements FlowInterface
 
     public function fn(FlowInterface $flow): FlowInterface
     {
-        if($this->fnFlow) {
+        if ($this->fnFlow) {
             $this->fnFlow->fn($flow);
         } else {
             $this->fnFlow = $flow;